Refactor friend and profile pages
